% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawPRcurve.R
\name{Draw_PR_curve}
\alias{Draw_PR_curve}
\title{Plot Precision-Recall Curve}
\usage{
Draw_PR_curve(pred_tensor, label_tensor)
}
\arguments{
\item{pred_tensor}{1D torch tensor of predictions (real-valued scores)}

\item{label_tensor}{1D torch tensor of labels (-1 = negative, 1 = positive)}
}
\value{
A ggplot2 object displaying the PR curve.
}
\description{
Computes and visualizes the precision-recall (PR) curve.
}
\seealso{
[PR_curve()]
}
