% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifPRaum.R
\name{MeasureClassifPRAUM}
\alias{MeasureClassifPRAUM}
\title{Custom mlr3 Measure: PR AUM}
\format{
[R6::R6Class] object.
}
\description{
An [mlr3::MeasureClassif] implementation that evaluates binary classifiers
using the PR AUM metric (Area Under Minimum(FDR,FNR)).
}
\details{
This measure is designed for imbalanced binary classification. It treats
labels as `-1` (negative class) and `+1` (positive class), converts
predictions into torch tensors, and computes the AUM score via
[PR_AUM()]. The measure is minimized, with values closer to 0
indicating better performance.

The `initialize()` method sets up the measure with default properties.
Users should call `MeasureClassifPRAUM$new()` rather than `initialize()` directly.
}
\references{
Original implementation ideas adapted from:
\itemize{
  \item \url{https://github.com/OGuenoun/R-AUM_Multiclass/blob/main/AUM_comparison.r}
  \item \url{https://github.com/tdhock/imbalanced-paper/blob/main/2025-07-30-conv-linear-prop0.01.R}
}
}
\seealso{
[PR_AUM()], [PR_curve()]
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{MeasureClassifPRAUM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureClassifPRAUM-new}{\code{MeasureClassifPRAUM$new()}}
\item \href{#method-MeasureClassifPRAUM-clone}{\code{MeasureClassifPRAUM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifPRAUM-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifPRAUM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifPRAUM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifPRAUM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifPRAUM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifPRAUM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
