% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRcurve.R
\name{PR_curve}
\alias{PR_curve}
\title{Precision-Recall Curve Statistics}
\usage{
PR_curve(pred_tensor, label_tensor)
}
\arguments{
\item{pred_tensor}{A 1D [torch::torch_tensor] of real-valued prediction
scores (higher values indicate stronger evidence for the positive class).}

\item{label_tensor}{A 1D [torch::torch_tensor] of true class labels,
encoded as `-1` for the negative class and `+1` for the positive class.}
}
\value{
A named list of [torch::torch_tensor] objects:
\itemize{
  \item `recall` - recall at each threshold
  \item `precision` - precision at each threshold
  \item `FNR` - false negative rate
  \item `FDR` - false discovery rate
  \item `"min(FDR,FNR)"` - minimum of FDR and FNR
  \item `min_constant` - left endpoints of the  constant added to predicted score
  \item `max_constant` - right endpoints of the  constant added to predicted score
}
}
\description{
Computes the precision-recall curve and related statistics from predicted
probabilities and binary labels using `torch` tensors.
}
\details{
This function computes the recall, precision, False Negative Rate (FNR), 
False Discovery Rate (FDR) but also the minimum between the FNR and the FDR. 
The min_constant and max_constant give the range of constants which result in
the corresponding error values.
}
